requirements.txt
itsdangerous==1.1.0
Jinja2==3.0.0
MarkupSafe==2.0.0
Werkzeug==1.0.1
numpy
scipy>=1.6.2
scikit-learn==1.2.2
matplotlib>=3.3.4
pandas>=1.2.4
streamlit
gunicorn==21.2.0
pickle-mixin
seaborn
matplotlib
dill
sklearn_pandas
pyyaml
mlflow
dvc
#-e .